Lint java

+++ src/main/java/ml/duncte123/skybot/objects/user/ConsoleUser.java
@@ -19,11 +19,10 @@
 package ml.duncte123.skybot.objects.user;
 
 import ml.duncte123.skybot.Author;
-import ml.duncte123.skybot.objects.user.FakeUser;
 
 @Author(nickname = "duncte123", author = "Duncan Sterken")
 public class ConsoleUser extends FakeUser {
     public ConsoleUser() {
-        super("DuncteBot", 210363111729790977L, (short) 1758);
+        super("DuncteBot", 210363111729790977L, 1758);
     }
 }
